feat: dead code removal, profling foudnation improvements, addressing code smells in reference db implementation
