--- src/compression/compression_scheduler.cpp
@@ -1,5 +1,0 @@
-// Compression scheduler implementation is included in compression_policy.cpp
-// This file exists to satisfy the CMakeLists.txt entry
-
-// The CompressionScheduler::Impl class is defined in compression_policy.cpp
-// to keep related functionality together and avoid circular dependencies
