--- src/daemon/components/WalMetricsProvider.cpp
@@ -1,7 +1,0 @@
-#include <yams/daemon/components/WalMetricsProvider.h>
-
-namespace yams::daemon {
-
-// All logic in header; this TU ensures the provider participates in yams_daemon linkage
-
-} // namespace yams::daemon
