+++ src/vector/turboquant.cpp
@@ -993,197 +993,6 @@ TurboQuantMSE::FittedTurboQuantModel TurboQuantMSE::fittedModel() const {
     return m;
 }
 
-// =============================================================================
-// TurboQuantProd Implementation (Inner Product Mode)
-// =============================================================================
-
-TurboQuantProd::TurboQuantProd(const TurboQuantConfig& config)
-    : config_(config), mse_quantizer_([&config]() {
-          TurboQuantConfig mse_config = config;
-          if (config.bits_per_channel > 1) {
-              mse_config.bits_per_channel = config.bits_per_channel - 1;
-          } else {
-              mse_config.bits_per_channel = 1;
-          }
-          return TurboQuantMSE(mse_config);
-      }()) {
-    // QJL parameters for inner product estimation
-    qjl_m_ = config.qjl_m;
-    if (qjl_m_ == 0) {
-        qjl_m_ = config.dimension / 4;
-    }
-
-    // Generate QJL random projection matrix
-    generateQJLCMatrix();
-}
-
-void TurboQuantProd::generateQJLCMatrix() {
-    const size_t d = config_.dimension;
-    qjl_matrix_.resize(qjl_m_ * d);
-
-    std::mt19937 rng(static_cast<uint32_t>(config_.seed + 12345));
-    std::normal_distribution<float> dist(0.0f, 1.0f / std::sqrt(static_cast<float>(d)));
-
-    for (size_t i = 0; i < qjl_m_ * d; ++i) {
-        qjl_matrix_[i] = dist(rng);
-    }
-}
-
-TurboQuantEncoded TurboQuantProd::encode(const std::vector<float>& vector) {
-    // First stage: MSE quantization
-    auto mse_indices = mse_quantizer_.encode(vector);
-
-    // Compute residual: r = x - x_tilde
-    auto reconstructed = mse_quantizer_.decode(mse_indices);
-    std::vector<float> residual(vector.size());
-    float residual_norm_sq = 0.0f;
-    for (size_t i = 0; i < config_.dimension; ++i) {
-        residual[i] = vector[i] - reconstructed[i];
-        residual_norm_sq += residual[i] * residual[i];
-    }
-
-    // Second stage: QJL on residual
-    std::vector<int8_t> qjl_signs(qjl_m_);
-    for (size_t m = 0; m < qjl_m_; ++m) {
-        float sum = 0.0f;
-        for (size_t i = 0; i < config_.dimension; ++i) {
-            sum += qjl_matrix_[m * config_.dimension + i] * residual[i];
-        }
-        qjl_signs[m] = (sum >= 0.0f) ? 1 : -1;
-    }
-
-    return {std::move(mse_indices), std::move(qjl_signs), residual_norm_sq};
-}
-
-/**
- * Estimate inner product between two vectors using TurboQuant encoding.
- *
- * Conservative blend correction: without residual norms we cannot apply the full
- * analytical correction. Instead we use sign agreement as a confidence signal.
- *
- * @return Approximate inner product with conservative QJL correction
- */
-float TurboQuantProd::estimateInnerProduct(
-    const std::pair<std::vector<uint8_t>, std::vector<int8_t>>& enc1,
-    const std::pair<std::vector<uint8_t>, std::vector<int8_t>>& enc2) {
-    // Decode first vector from MSE indices
-    auto x1 = mse_quantizer_.decode(enc1.first);
-
-    // Decode second vector from MSE indices
-    auto x2 = mse_quantizer_.decode(enc2.first);
-
-    // MSE-decoded inner product (base estimate)
-    float dot_mse = 0.0f;
-    for (size_t i = 0; i < config_.dimension; ++i) {
-        dot_mse += x1[i] * x2[i];
-    }
-
-    // QJL correction: measure sign agreement rate and apply confidence adjustment
-    const size_t m = enc1.second.size();
-    if (m == 0 || enc1.second.size() != enc2.second.size()) {
-        // No QJL stage or mismatched sizes — fall back to MSE-only estimate
-        return dot_mse;
-    }
-
-    size_t matching_signs = 0;
-    for (size_t i = 0; i < m; ++i) {
-        if (enc1.second[i] == enc2.second[i]) {
-            matching_signs++;
-        }
-    }
-    float gamma = static_cast<float>(matching_signs) / static_cast<float>(m);
-
-    // For independent Gaussian projections the expected agreement is 0.5.
-    // The deviation (gamma - 0.5) tells us about residual correlation.
-    // We apply a mild multiplicative correction: blend factor 0.25 means
-    // a perfect agreement (gamma=1) boosts the estimate by 12.5%.
-    // This is conservative — it nudges rather than over-corrects.
-    constexpr float kQjlBlendFactor = 0.25f;
-    float correction = 1.0f + kQjlBlendFactor * (gamma - 0.5f) * 2.0f;
-    return dot_mse * correction;
-}
-
-/**
- * Full QJL-corrected inner product estimation using residual norms.
- *
- * NOTE (2026-03-25): Benchmark results show this approach is WORSE than decoded-only
- * for random vectors. For d=384, b=4:
- *   - Decoded MAE: 0.042
- *   - QJL-corrected MAE: 0.357
- *   - QJL correction adds noise because residual norms are small and residuals are
- *     near-orthogonal for random vectors (ρ_res ≈ 0), making the sign agreement
- *     rate a noisy proxy for a near-zero quantity.
- *
- * This method is retained for future exploration on structured/correlated data where
- * residuals are NOT independent. For production use, prefer estimateInnerProduct() which
- * uses a conservative blend that doesn't actively worsen the estimate.
- *
- * QJL correction theory (corrected 2026-03-25):
- *   For random Gaussian projections S_ij ~ N(0, 1/d), each projected component
- *   (S·r)_m ~ N(0, ||r||²) where r is the residual (post-MSE quantization error).
- *
- *   Joint Normal Lemma:
- *     P(both positive) = 1/2 + (1/π)·arcsin(ρ)   [Feller Vol.2, Ch. III.4]
- *     E[sgn·sgn] = 2·P(both same) - 1 = (2/π)·arcsin(ρ)
- *
- *   Key derivation:
- *     γ = (#matching) / m  ≈  P(both same sign)         // observed agreement rate
- *     E[sgn·sgn] = 2·γ - 1                            // sign product expectation
- *     2·γ - 1 = (2/π)·arcsin(ρ)                      // equate
- *     arcsin(ρ) = π·γ - π/2
- *     ρ_res = sin(π·γ - π/2) = -cos(π·γ)             // correct inverse
- *
- *   The full QJL-corrected inner product:
- *     est_ip = dot(MSE) + ρ_res · √(r1_norm_sq · r2_norm_sq)
- *
- * @return Approximate inner product
- */
-float TurboQuantProd::estimateInnerProductFull(const TurboQuantEncoded& enc1,
-                                               const TurboQuantEncoded& enc2) {
-    // MSE-decoded inner product
-    auto x1 = mse_quantizer_.decode(enc1.mse_indices);
-    auto x2 = mse_quantizer_.decode(enc2.mse_indices);
-    float dot_mse = 0.0f;
-    for (size_t i = 0; i < config_.dimension; ++i) {
-        dot_mse += x1[i] * x2[i];
-    }
-
-    // QJL correction using sign agreement rate
-    const size_t m = enc1.qjl_signs.size();
-    if (m == 0 || enc1.qjl_signs.size() != enc2.qjl_signs.size()) {
-        return dot_mse;
-    }
-
-    size_t matching_signs = 0;
-    for (size_t i = 0; i < m; ++i) {
-        if (enc1.qjl_signs[i] == enc2.qjl_signs[i]) {
-            matching_signs++;
-        }
-    }
-    float gamma = static_cast<float>(matching_signs) / static_cast<float>(m);
-
-    // Correct inversion of the Joint Normal Lemma:
-    //   arcsin(ρ) = π·γ - π/2   →   ρ = sin(π·γ - π/2) = -cos(π·γ)
-    // Using -cos(π·γ) directly avoids numerical issues at the boundaries.
-    float rho_res = -std::cos(static_cast<float>(M_PI) * gamma);
-
-    // Residual dot product contribution: ρ_res · ||r1|| · ||r2||
-    float residual_contribution =
-        rho_res * std::sqrt(enc1.residual_norm_sq * enc2.residual_norm_sq);
-
-    return dot_mse + residual_contribution;
-}
-
-std::vector<float> TurboQuantProd::decode(const std::vector<uint8_t>& mse_indices) {
-    return mse_quantizer_.decode(mse_indices);
-}
-
-double computeDistortionLowerBound(size_t d, size_t total_bits) {
-    // Shannon distortion-rate lower bound: D >= (d/2π) * 2^(-2B/d)
-    double B = static_cast<double>(total_bits) / static_cast<double>(d);
-    return (static_cast<double>(d) / (2.0 * M_PI)) * std::pow(2.0, -2.0 * B);
-}
-
 float sampleBeta(std::mt19937& rng, float alpha, float beta) {
     // Using gamma distribution method: X = X1/(X1+X2) where X1~Gamma(α,1), X2~Gamma(β,1)
     std::gamma_distribution<float> gamma_alpha(alpha, 1.0f);
