+++ tests/benchmarks/meson.build
@@ -193,8 +193,8 @@ if benchmark_dep.found()
 
   retrieval_quality_bench = executable('retrieval_quality_bench',
     files(
-      '../../src/search/benchmarks/retrieval_quality_bench.cpp',
-      '../../src/search/benchmarks/retrieval_benchmark_support.cpp',
+      'search/retrieval_quality_bench.cpp',
+      'search/retrieval_benchmark_support.cpp',
     ),
     include_directories: incs,
     dependencies: retrieval_quality_bench_deps,
@@ -208,8 +208,8 @@ if benchmark_dep.found()
   # Fusion strategy experiment - compares different fusion approaches
   fusion_experiment = executable('fusion_strategy_experiment',
     files(
-      '../../src/search/benchmarks/fusion_strategy_experiment.cpp',
-      '../../src/search/benchmarks/retrieval_benchmark_support.cpp',
+      'search/fusion_strategy_experiment.cpp',
+      'search/retrieval_benchmark_support.cpp',
     ),
     include_directories: incs,
     dependencies: retrieval_quality_bench_deps,
@@ -223,7 +223,7 @@ if benchmark_dep.found()
   ingestion_e2e_bench_deps = search_bench_deps + [test_deps]
 
   ingestion_e2e_bench = executable('ingestion_e2e_bench',
-    files('../../src/search/benchmarks/ingestion_e2e_bench.cpp'),
+    files('search/ingestion_e2e_bench.cpp'),
     include_directories: incs,
     dependencies: ingestion_e2e_bench_deps,
     cpp_args: ['-DYAMS_TESTING=1'] + tracy_args,
